+++ src/Unreal.Core/Models/ExternalData.cs
@@ -11,6 +11,6 @@ namespace Unreal.Core.Models
     {
         public uint NetGUID { get; init; }
         public FArchive Archive { get; init; }
-        public int TimeSeconds { get; init; }
+        public float TimeSeconds { get; init; }
     }
 }
